third_party/WebKit/LayoutTests/typedcssom/inlinestyle/inlineStylePropertyMap_iteration.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<div id="testElement"></div>

<script>

test(function() {
  testElement.style = "";

  var entryIterator = testElement.styleMap.entries();
  var keyIterator = testElement.styleMap.keys();
  var valueIterator = testElement.styleMap.values();

  assert_true(entryIterator.next().done);
  assert_true(keyIterator.next().done);
  assert_true(valueIterator.next().done);
}, "Iteration over empty StyleMap returns empty iterator");

test(function() {
  testElement.style = "width: 50px";

  var iterator = testElement.styleMap.entries();
  var entry = iterator.next();
  assert_false(entry.done);
  // Should only be one entry.
  assert_true(iterator.next().done);

  assert_equals(entry.value[0], 'width');
  assert_true(entry.value[1] instanceof CSSSimpleLength);
  assert_equals(entry.value[1].cssText, '50px');
}, "Iterator for single entry returns iterator with a single value");

test(function() {
  testElement.style = "width: 60px";

  var iterator = testElement.styleMap.keys();
  var entry = iterator.next();
  assert_false(entry.done);
  // Should only be one entry.
  assert_true(iterator.next().done);

  assert_equals(entry.value, 'width');
}, "Iterator for single key returns iterator with a single value");

test(function() {
  testElement.style = "width: 70px";

  var iterator = testElement.styleMap.values();
  var entry = iterator.next();
  assert_false(entry.done);
  // Should only be one entry.
  assert_true(iterator.next().done);

  assert_true(entry.value instanceof CSSSimpleLength);
  assert_equals(entry.value.cssText, '70px');
}, "Iterator for single value returns iterator with a single value");

test(function() {
  testElement.style = "border: 5px solid lightcoral";

  var entries = {};
  var numEntries = 0;
  for (let value of testElement.styleMap.entries()) {
    numEntries++;
    entries[value[0]] = value[1];
  }
  assert_equals(numEntries, 17);

  assert_equals(entries['border-top-width'].cssText, '5px');
  assert_equals(entries['border-right-width'].cssText, '5px');
  assert_equals(entries['border-bottom-width'].cssText, '5px');
  assert_equals(entries['border-left-width'].cssText, '5px');

  assert_equals(entries['border-top-style'].cssText, 'solid');
  assert_equals(entries['border-right-style'].cssText, 'solid');
  assert_equals(entries['border-bottom-style'].cssText, 'solid');
  assert_equals(entries['border-left-style'].cssText, 'solid');

  assert_equals(entries['border-top-color'].cssText, 'lightcoral');
  assert_equals(entries['border-right-color'].cssText, 'lightcoral');
  assert_equals(entries['border-bottom-color'].cssText, 'lightcoral');
  assert_equals(entries['border-left-color'].cssText, 'lightcoral');

  assert_equals(entries['border-image-source'].cssText, 'initial');
  assert_equals(entries['border-image-slice'].cssText, 'initial');
  assert_equals(entries['border-image-width'].cssText, 'initial');
  assert_equals(entries['border-image-outset'].cssText, 'initial');
  assert_equals(entries['border-image-repeat'].cssText, 'initial');
}, "Iterating entries over border element expansion");

test(function() {
  testElement.style = "background-color: var(--bg-color);";
  var entries = [...testElement.styleMap.values()];

  assert_equals(entries.length, 1);
  assert_equals(entries[0].constructor, CSSUnparsedValue);
  assert_equals(entries[0].cssText, '--bg-color');
  assert_equals([...entries[0]].length, 1);
  assert_equals([...entries[0]][0].variable, "--bg-color");
  assert_equals([...entries[0]][0].fallback, null);
}, "Variable values come out as CSSUnparsedValues");

test(function() {
  testElement.style = '';
  testElement.style.setProperty('--my-custom-property', '5px');

  var entries = [...testElement.styleMap.entries()];
  assert_equals(entries.length, 1);
  var propertyAndValue = entries[0];
  assert_equals(propertyAndValue.length, 2);

  assert_equals(propertyAndValue[0], '--my-custom-property');
  assert_equals(propertyAndValue[1].constructor, CSSStyleValue);
  assert_equals(propertyAndValue[1].cssText, '5px');
}, "Custom properties set on the element come out as CSSStyleValues");

test(function() {
  testElement.style.cssText = "@apply --foo";

  var entries = [...testElement.styleMap.entries()];
  assert_equals(entries.length, 1);
  var propertyAndValue = entries[0];
  assert_equals(propertyAndValue.length, 2);

  assert_equals(propertyAndValue[0], '@apply');
  assert_equals(propertyAndValue[1].constructor, CSSStyleValue);
  assert_equals(propertyAndValue[1].cssText, '--foo');
}, "@apply rules come out as CSSStyleValues");

</script>
